Fix problems uncovered while building the parser
This adds the __repr__() method to the Instruction class (now renamed
to AbstractInstruction) for better debugging of the produced
instructions.  Also adds the Instructions subclass of
AbstractInstruction (the reason for the rename) to allow grouping of
sequences of instructions into a single instruction; this was
necessary to deal with problems in the parser.  Finally, the fold()
methods were updated to return Constant or Instructions instances,
depending on whether folding was done or not; previously fold() would
pass the actual Constant object to the operation, which did not work,
and would return a list of instructions, which confused the parser.