Create, document, and test Authorization
The Authorization object is the results of an authorization check.  It
is constructed by the policies.instructions.set_authz instruction, and
updated by the policies.instructions.AuthorizationAttr instructions.
It has a boolean value indicating whether authorization succeeded or
not, and allows arbitrary authorization attributes to be queried
(defaulting to None if the value hasn't been set to a default on the
rule or by a parsed authorization rule).