A pyparsing parser for authorization rules
This contains parse_rule(), which parses the text of an authorization
rule and returns a policies.instructions.Instructions object which
implements the desired rule.  If the rule does not parse, parse_rule()
will either log the error (the default) or allow the parsing exception
to bubble up to the caller.

Although parse_rule() is the only symbol exported, a great number of
variables exist, which contain pyparsing expressions.  These
expressions include a hand-written expression parser; we do not use
pyparsing.infixNotation() because the expression syntax includes some
operations that cannot be expressed through its interface: attribute
access, item access, and function call.