Create TrinaryOperator
It was discovered that the trinary_op did not properly fold constants,
because it uses the default fold() implementation of Operator, which
requires all arguments to be constant.  For trinary, only the
conditional needs to be constant, so we create a new TrinaryOperator
class which overrides fold() to implement the appropriate logic.