Begin sketching in rules
This defines a single Rule class, which describes a single rule.  Objects
of this class are immutable; to replace the rule, you create a new Rule
instance.  The Rule class makes the compiled instructions available
through the 'instructions' property, which performs lazy compilation to
amortize the compilation cost, and caches the result for efficiency.

This will be followed up by a RuleSet class, which will allow for the
predeclaration of rules and evaluation of a given rule, as well as the
RuleContext class, which is the ctxt object expected by
AbstractInstruction.__call__().