Add @want_context
Some functions meant to be callable from the policy language may want
to manipulate the context directly.  This adds a @want_context
decorator that marks such functions.  If used on a function, the
context will be passed as the first argument to the function, with
the remaining arguments being popped off the stack, and it is expected
that the function itself will perform any requisite stack manipulation.