Add the rule() function
This adds the special rule() function to the policy language.  The
rule() function allows the result of evaluating another rule to be
inserted into the evaluation context stack.  The evaluation of the
other rule is stopped at the set_authz instruction, so no
authorization attributes will be set.  The result of evaluating a
rule is cached for efficiency, allowing the same rule to be used
both in the main policy and in the authorization attributes without
having to evaluate the rule multiple times.  The caching is only for
the duration of the Policy.evaluate() method.