Add recursion detection
Now that rules can reference other rules, there's the possibility that
a recursive rule evaluation could occur.  This adds a simple safeguard
to prevent that: when a rule name is pushed onto the context with
push_rule(), a check is made to ensure the rule name is not already on
the rule name stack, and an exception is raised if it is.