Reimplement some operators to short-circuit
The "and", "or", and trinary operators need to short-circuit; that is,
they should only evaluate some of their operands, rather than always
evaluating all of them.  This change uses the new jump instructions to
implement this.  It also creates a distinction between operators
(which implement a fold() method) and instructions.